ud: Move test files around
This separates voice tests from admin in preparation for
adding another teams file.
