+++ unmnemonic_devices_vrs/tests/admin/fixtures
@@ -1,0 +1,1 @@
+../fixtures
