+++ unmnemonic_devices_vrs/tests/admin/main.rs
@@ -1,0 +1,1 @@
+mod calls;
