+++ unmnemonic_devices_vrs/tests/fixtures/destinations.sql
