+++ unmnemonic_devices_vrs/tests/fixtures/meetings.sql
