+++ unmnemonic_devices_vrs/tests/fixtures/recordings-prompts-pure-monitoring.sql
