+++ unmnemonic_devices_vrs/tests/fixtures/regions.sql
