+++ unmnemonic_devices_vrs/tests/fixtures/schema.sql
