+++ unmnemonic_devices_vrs/tests/fixtures/settings-ending.sql
