+++ unmnemonic_devices_vrs/tests/fixtures/settings.sql
