+++ unmnemonic_devices_vrs/tests/fixtures/teams.sql
