+++ unmnemonic_devices_vrs/tests/voice/fixtures
@@ -1,0 +1,1 @@
+../fixtures
