+++ unmnemonic_devices_vrs/tests/voice/recordings.rs
@@ -1,4 +1,8 @@
-use crate::helpers::{get, post, RedirectTo};
+mod helpers {
+    include!("../helpers.rs");
+}
+use helpers::{get, post, RedirectTo};
+
 use select::{document::Document, predicate::Name};
 use serde::Serialize;
 use speculoos::prelude::*;
